+++ src/configuration.ts
@@ -3,25 +3,42 @@ import { EXTENSION_PREFIX } from './constants';
 
 export class Configuration {
   /**
-   * Get a {@link EXTENSION_PREFIX `formatto`}-scoped workspace configuration object.
+   * Get a {@link EXTENSION_PREFIX `formatto`} configuration object scoped by {@link scope}.
    */
-  protected get configuration(): vscode.WorkspaceConfiguration {
-    return vscode.workspace.getConfiguration(EXTENSION_PREFIX);
+  private getConfiguration(scope?: vscode.Uri): vscode.WorkspaceConfiguration {
+    return vscode.workspace.getConfiguration(EXTENSION_PREFIX, scope);
+  }
+
+  /**
+   * Return an unscoped value from {@link EXTENSION_PREFIX} configuration.
+   */
+  private getGlobal<T>(section: string): T | undefined;
+  private getGlobal<T>(section: string, defaultValue: T): T;
+  private getGlobal<T>(section: string, defaultValue?: T): T | undefined {
+    return this.getConfiguration(undefined).get(section, defaultValue);
+  }
+
+  public get enableRangeFormatting(): boolean {
+    return this.getGlobal('enableRangeFormatting', false);
   }
 
   /**
    * Get `rubyfmt` path.
    */
-  public getRubyFmtPath(): string {
-    return this.configuration.get('rubyfmtPath', 'rubyfmt');
+  public getRubyfmtPath(scope?: vscode.Uri): string {
+    return this.getValue(scope, 'rubyfmtPath', 'rubyfmt');
   }
 
   /**
-   * Return a value from {@link EXTENSION_PREFIX} configuration.
+   * Return a {@link scope}-scoped value from {@link EXTENSION_PREFIX} configuration.
    */
-  public getValue<T>(section: string): T | undefined;
-  public getValue<T>(section: string, defaultValue: T): T;
-  public getValue<T>(section: string, defaultValue?: T): T | undefined {
-    return this.configuration.get(section, defaultValue);
+  private getValue<T>(scope: vscode.Uri | undefined, section: string): T | undefined;
+  private getValue<T>(scope: vscode.Uri | undefined, section: string, defaultValue: T): T;
+  private getValue<T>(
+    scope: vscode.Uri | undefined,
+    section: string,
+    defaultValue?: T,
+  ): T | undefined {
+    return this.getConfiguration(scope).get(section, defaultValue);
   }
 }