+++ src/rubyfmt.ts
@@ -2,14 +2,26 @@ import * as vscode from 'vscode';
 import { spawn } from 'child_process';
 import { ExtensionContext } from './extensionContext';
 
-export async function rubyfmt(
+export function formatDocument(
   context: ExtensionContext,
-  source: string,
+  document: vscode.TextDocument,
   token: vscode.CancellationToken,
 ): Promise<string> {
-  const rubyfmtPath = context.configuration.getRubyFmtPath();
+  return formatText(context, document.getText(), document.uri, token);
+}
+
+export async function formatText(
+  context: ExtensionContext,
+  text: string,
+  uri: vscode.Uri,
+  token: vscode.CancellationToken,
+): Promise<string> {
+  const rubyfmtPath = context.configuration.getRubyfmtPath(uri);
+  const workspaceFolder = uri && vscode.workspace.getWorkspaceFolder(uri)?.uri;
+
   const formattedSource = new Promise<string>((resolve, reject) => {
     const child = spawn(rubyfmtPath, [], {
+      cwd: workspaceFolder?.fsPath,
       stdio: ['pipe', 'pipe', 'pipe'],
     });
 
@@ -37,14 +49,29 @@ export async function rubyfmt(
       }
 
       if (code !== 0) {
-        reject(new Error(`rubyfmt exited with code ${code}: ${stderr}`));
+        reject(new Error(`rubyfmt exited with code ${code}: ${stderr.trim()}`));
       } else {
         resolve(stdout);
       }
     });
 
-    child.stdin.write(source);
+    child.stdin.write(text);
     child.stdin.end();
   });
+  return formattedSource;
+}
+
+export async function tryFormatDocument(
+  context: ExtensionContext,
+  document: vscode.TextDocument,
+  token: vscode.CancellationToken,
+): Promise<string | undefined> {
+  let formattedSource: string | undefined;
+  try {
+    formattedSource = await formatDocument(context, document, token);
+  } catch (e) {
+    const message = e instanceof Error ? e.message : (e?.toString() ?? 'Failed to format document');
+    context.log.error(message, vscode.workspace.asRelativePath(document.uri));
+  }
   return formattedSource;
 }